third_party/blink/web_tests/wpt_internal/private-aggregation/shared-storage-sends-report.https.html

<!doctype html>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>

<body>
<script>
'use strict';

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: true, contributions: [{bucket: 1n, value: 2}]};
  await sharedStorage.run("contribute-to-histogram", {data, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/true,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/ONE_CONTRIBUTION_EXAMPLE_PAYLOAD);

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage")
  assert_equals(debug_reports.length, 1);

  verifyReportsIdenticalExceptPayload(report, JSON.parse(debug_reports[0]));

}, 'run() that calls Private Aggregation with one contribution');

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: true,
                contributions: [{bucket: 1n, value: 2}, {bucket: 3n, value: 4}]};

  await sharedStorage.run("contribute-to-histogram", {data, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/true,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/MULTIPLE_CONTRIBUTIONS_EXAMPLE_PAYLOAD);

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage")
  assert_equals(debug_reports.length, 1);

  verifyReportsIdenticalExceptPayload(report, JSON.parse(debug_reports[0]));
}, 'run() that calls Private Aggregation with multiple contributions');

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: true,
                enableDebugModeArgs: {debugKey: 1234n},
                contributions: [{bucket: 1n, value: 2}]};

  await sharedStorage.run("contribute-to-histogram", {data, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/true,
               /*debug_key=*/"1234",
               /*expected_cleartext_payload=*/ONE_CONTRIBUTION_EXAMPLE_PAYLOAD);

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage")
  assert_equals(debug_reports.length, 1);

  verifyReportsIdenticalExceptPayload(report, JSON.parse(debug_reports[0]));
}, 'run() that calls Private Aggregation with debug key');


private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: true,
                enableDebugModeArgs: {debugKey: 1234n},
                contributions: [{bucket: 1n, value: 2}, {bucket: 3n, value: 4}]};

  await sharedStorage.run("contribute-to-histogram", {data, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/true,
               /*debug_key=*/"1234",
               /*expected_cleartext_payload=*/MULTIPLE_CONTRIBUTIONS_EXAMPLE_PAYLOAD);

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage")
  assert_equals(debug_reports.length, 1);

  verifyReportsIdenticalExceptPayload(report, JSON.parse(debug_reports[0]));
}, 'run() that calls Private Aggregation with multiple contributions and a debug key');

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: false,
                contributions: [{bucket: 1n, value: 2}]};

  await sharedStorage.run("contribute-to-histogram", {data, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);

  // Note that the payload cannot be tested as debug mode is disabled.
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/false,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/undefined);

  // We use a short timeout as the previous poll should've waited long enough.
  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/50)
  assert_equals(debug_reports, null);
}, 'run() that calls Private Aggregation without debug mode');

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugModeAfterOp: true,
                enableDebugModeArgs: {debugKey: 1234n},
                contributions: [{bucket: 1n, value: 2}]};

  await sharedStorage.run("contribute-to-histogram", {data, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/true,
               /*debug_key=*/"1234",
               /*expected_cleartext_payload=*/ONE_CONTRIBUTION_EXAMPLE_PAYLOAD);

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage")
  assert_equals(debug_reports.length, 1);

  verifyReportsIdenticalExceptPayload(report, JSON.parse(debug_reports[0]));
}, 'run() that calls enableDebugMode() after contributeToHistogram()');


private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data_1 = {enableDebugMode: true,
                  contributions: []};
  const data_2 = {contributions: [{bucket: 1n, value: 2}]};

  await sharedStorage.run("contribute-to-histogram", {data: data_1, keepAlive: true});
  await sharedStorage.run("contribute-to-histogram", {data: data_2, keepAlive: true});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage")
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/false,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/undefined);

  // We use a short timeout as the previous poll should've waited long enough.
  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/50)
  assert_equals(debug_reports, null);
}, 'calling enableDebugMode() in a different operation doesn\'t apply');


private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data_1 = {enableDebugMode: true,
                  contributions: [{bucket: 1n, value: 2}]};
  const data_2 = {enableDebugMode: true,
                  contributions: [{bucket: 1n, value: 2}, {bucket: 3n, value: 4}]};

  await sharedStorage.run("contribute-to-histogram", {data: data_1, keepAlive: true});
  await sharedStorage.run("contribute-to-histogram", {data: data_2, keepAlive: true});

  // We don't verify the reports as they could arrive in a different order.
  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage", /*wait_for=*/2);
  assert_equals(reports.length, 2);

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage",
      /*wait_for=*/2);
  assert_equals(debug_reports.length, 2);
}, 'reports from multiple operations aren\'t batched');

</script>
</body>
